5_retrieval_augmented_generation(rag)/requirements.txt
langchain
langchain-openai
openai
langchain-google-genai
cohere
faiss-cpu
streamlit
python-dotenv
llama-index
pypdf
chromadb
beautifulsoup4
matplotlib
rank_bm25